FLEX-232 : Spring Flex should work with Spring 4.x
Previous version was compatible with Spring 3.0.x,
but not with Spring 3.1.x, 3.2.x or 4.x.

HibernateProxyConverter is the main class broken if we try to upgrade
Spring. Previous version was using PropertyTypeDescriptor, but this
class has been removed since Spring 3.1.

The alternative is to use TypeDescriptor#getSource() in order to detect
if the target is a field or a property. This is the solution implemented
by this commit. Since this API is not available in Spring 3.1.x or Spring
3.2.x, this commit introduces Spring Framework 4.x and Java 6 as
minimum requirements for Spring Flex 1.6.x.